split + rename
